Pack200 - minor bug fixes
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/classlib/trunk@763569 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/pack200/src/main/java/org/apache/harmony/pack200/ClassBands.java
@@ -947,7 +947,7 @@ public class ClassBands extends BandSet {
 
     public void addAnnotationDefault(List nameRU, List t, List values, List caseArrayN, List nestTypeRS, List nestNameRU, List nestPairN) {
         method_AD_bands.addAnnotation(null, nameRU, t, values, caseArrayN, nestTypeRS, nestNameRU, nestPairN);
-        Integer flag = (Integer) tempMethodFlags.remove(tempMethodFlags.size() - 1);
-        tempMethodFlags.add(new Integer(flag.intValue() | (1<<25)));
+        Long flag = (Long) tempMethodFlags.remove(tempMethodFlags.size() - 1);
+        tempMethodFlags.add(new Long(flag.longValue() | (1<<25)));
     }
 }
